% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement_ads.R
\name{agreement_ads}
\alias{agreement_ads}
\title{Agreement's coefficients adjusted}
\usage{
agreement_ads(nrt, ni, na, data, NC = 4, NI = 100, data_prior = NULL,
  prior, index)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.
It can be a `as_BCM` object.}
}
\value{
A dataframe containing the couples compared (first column),
    and the SPA value (second column).
}
\description{
Thi function calculates different agreement's coefficients
between n-couples of raters
}
\examples{
data(toyEx)

## Compute SPA only between rater 1 and 3
agreement_ads(3,12,8,data=toyEx,prior="WI",index="Kappa")
}
\seealso{
`as_BCM`, `toyEx`
}
