% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMM_WIP.R
\name{BMM_WIP}
\alias{BMM_WIP}
\title{Belonging Measure Matrix (BMM) estimation-weakly informative prior}
\usage{
BMM_WIP(nrt, ni, na, data, NC = 4, NI = 100)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.
It can be a `as_BCM` object.}

\item{NC}{Number of chains. Default is equal to four.}

\item{NI}{Number of iteration per chain. Default is equal to 100}
}
\value{
TO BE DEFINED...
}
\description{
This function estimates the Belonging Measure Matrix, starting from
the Belonging Threshold in case of a weakly informative prior
}
\examples{
data(toyEx)

## Estimate BTs for three rater with a weakly informative prior
BMM_WIP(3,12,8,data=toyEx)
}
\seealso{
`as_BCM`, `toyEx`
}
