% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BT_IP.R
\name{BT_IP}
\alias{BT_IP}
\title{Belonging Threshold (BT) estimation- informative prior}
\usage{
BT_IP(nrt, ni, na, data, NC = 4, NI = 100, data_prior)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.
It can be a `as_BCM` object.}

\item{NC}{Number of chains. Default is equal to four.}

\item{NI}{Number of iteration per chain. Default is equal to 100}

\item{data_prior}{A dataframe containing data representing the informative prior}
}
\value{
TO BE DEFINED...
}
\description{
This function estimates theBelonging Threshold for each rater
in case of an informative prior
}
\examples{
data(toyEx)

## Estimate BTs for three rater with a weakly informative prior
expertBMM<-expertBMM[,2:9]
BT_IP(3,12,8,data=toyEx,data_prior=expertBMM)
}
\seealso{
`as_BCM`, `toyEx`
}
