% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTE-package.R
\docType{package}
\name{BTE-package}
\alias{BTE-package}
\alias{BTE}
\title{The 'BTE' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.1. https://mc-stan.org
}
