% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kappa_BCM.R
\name{kappa_BCMs}
\alias{kappa_BCMs}
\title{Cohen's kappa between BCMs}
\usage{
kappa_BCMs(nrt, ni, na, data, whichRaters)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.
It can be a `as_BCM` object.}
}
\value{
A dataframe containing the couples compared (first column),
    and the kappa value (second column).
}
\description{
Thi function calculates the Cohen's kappa index
between n-couples of Boolean Classification Matrices
}
\examples{
data(toyEx)

## Compute SPA only between rater 1 and 3
kappa_BCMs(3,12,8,data=toyEx,whichRaters=c(1,3))
}
\seealso{
`as_BCM`, `toyEx`
}
