% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPA.R
\name{SPA}
\alias{SPA}
\title{Simple Percentage Agreement between BCMs}
\usage{
SPA(nrt, ni, na, data, whichRaters)
}
\arguments{
\item{nrt}{Number of raters.}

\item{ni}{Number of items.}

\item{na}{Number of symptoms/attributes investigated by items.}

\item{data}{Dataframe containing the original attribution.
It can be a `as_BCM` object.}
}
\value{
A dataframe containing the couples compared (first column),
    and the SPA value (second column).
}
\description{
Thi function calculates the simple percentage agreement index
between n-couples of Boolean Classification Matrices
}
\examples{
data(toyEx)

## Compute SPA only between rater 1 and 3
SPA(3,12,8,data=toyEx,whichRaters=c(1,3))
}
\seealso{
`as_BCM`, `toyEx`
}
